Move memoize, uniq, difference, flatten, *_return_values to theano.utils
